Module Lwt_log_core.Section

Sections

type t = section

val make : string -> section

make name creates a section with the given name. Two calls to make with the same name will return the same section object.

val name : section -> string

name section returns the name of section.

val main : section

The main section. It is the section used by default when no one is provided.

val level : section -> level

level section returns the logging level of section.

val set_level : section -> level -> unit

set_level section sets the logging level of the given section. Modifications of rules using add_rule won't affect the level of this section after this operation.

val reset_level : section -> unit

reset_level section resets the level of section to its default value, i.e. to the value obtained by applying rules.