Module Lwt_log_rules

Logging rules parsing

val rules : Stdlib.Lexing.lexbuf -> (string * string) list option

parse lexbuf returns the list of rules contained in lexbuf or None in case of parsing error